Extracted export modal
