+++ src/components/header/export-modal/index.css
@@ -1,0 +1,88 @@
+.export-content > p {
+  padding: 30px 0 0 0;
+  margin: 0;
+}
+
+.export-buttons {
+  display: flex;
+  flex-wrap: wrap;
+}
+
+.export-button {
+  min-width: 280px;
+  display: inline-flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: left;
+  font-size: 0.9em;
+  background: white;
+  border: 1px solid #ccc;
+  border-radius: 4px;
+  cursor: pointer;
+  user-select: none;
+  padding: 10px 20px;
+  margin-right: 40px;
+  width: 30%;
+  margin: 1%;
+}
+
+.export-button:hover {
+  background: #ddd;
+}
+
+.export-button > div {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+}
+
+.export-button span {
+  padding-left: 8px;
+  font-size: 1.6em;
+  line-height: 50px;
+}
+
+.export-button p {
+  color: #333;
+  line-height: 1.2;
+}
+
+.export-button p strong {
+  color: firebrick;
+}
+
+.export-button svg {
+  width: 30px;
+  height: 30px;
+}
+
+.export-button.disabled {
+  cursor: wait;
+  background: #eee;
+  filter: grayscale(80%);
+  opacity: 0.8;
+}
+
+.export-button .type-input-container {
+  margin-top: 10px;
+}
+
+.exported-url {
+  display: flex;
+}
+
+.exported-url span:first-child {
+  width: 100%;
+  overflow: hidden;
+  vertical-align: middle;
+  line-height: 25px;
+  white-space: nowrap;
+}
+
+.exported-url a::selection {
+  background: rgba(0, 0, 0, 0.15);
+}
+
+.user-notes p {
+  padding: 20px 0;
+}
