+++ src/components/header/export-modal/index.tsx
@@ -1,0 +1,15 @@
+import { connect } from 'react-redux';
+import { bindActionCreators, Dispatch } from 'redux';
+import * as EditorActions from '../../../actions/editor';
+import { State } from '../../../constants/default-state';
+import Renderer from './renderer';
+
+export function mapStateToProps(state: State, ownProps) {
+  return {
+    vegaLiteSpec: state.vegaLiteSpec,
+    vegaSpec: state.vegaSpec,
+    view: state.view,
+  };
+}
+
+export default connect(mapStateToProps)(Renderer);
